requirements.txt
flask==0.10.1
pymongo==2.7.2
Flask-Material==0.1.1
neo4j-driver==1.7.4
neo4j-doc-manager==0.1.2
mongo-connector==2.1
IMDbPY==6.7